Extract query into separate file
