+++ client/web/src/enterprise/insights/pages/dashboards/dashboard-page/components/add-insight-modal/index.ts
@@ -1,0 +1,4 @@
+export { AddInsightModal } from './AddInsightModal'
+export type { AddInsightModalProps } from './AddInsightModal'
+
+export { GET_ACCESSIBLE_INSIGHTS_LIST } from './query'
